# docs/swagger.yaml
definitions:
  bo.BaseOrgInfoBo:
    properties:
      orgId:
        type: integer
      orgName:
        type: string
      orgOwnerId:
        type: integer
      outOrgId:
        type: string
      sourceChannel:
        type: string
    type: object
...
  inner_req.AddAppReq:
    properties:
      appId:
        description: AppId
        type: integer
      orgId:
        description: OrgId
        type: integer
      userId:
        description: UserId
        type: integer
    type: object
  inner_req.AddOrgOutCollaborator:
    properties:
      orgId:
        type: integer
      userId:
        type: integer
    type: object
  inner_req.AddPkgReq:
    properties:
      orgId:
        description: OrgId
        type: integer
      pkgId:
        description: PkgId
        type: integer
      userId:
        description: UserId
        type: integer
    type: object
...
  inner_req.DeleteAppReq:
    properties:
      appId:
        description: AppId
        type: integer
      orgId:
        description: OrgId
        type: integer
      userId:
        description: UserId
        type: integer
    type: object
  inner_req.DeletePkgReq:
    properties:
      orgId:
        description: OrgId
        type: integer
      pkgId:
        description: PkgId
        type: integer
      userId:
        description: UserId
        type: integer
    type: object
  inner_req.DeptListByIdsInnerReq:
    properties:
      ids:
        description: Ids ID列表
        items:
          type: integer
        type: array
      orgId:
        description: OrgId 组织ID
        type: integer
    type: object
  inner_req.GetDeptUserIdsReq:
    properties:
      orgId:
        description: 组织id
        type: integer
    type: object
  inner_req.GetLeadersByDeptIdsReq:
    properties:
      deptIds:
        description: 部门id
        items:
          type: integer
        type: array
      orgId:
        description: 组织id
        type: integer
    type: object
...
  inner_req.GetMemberSimpleInfoReq:
    properties:
      orgId:
        description: 组织id
        type: integer
      type:
        description: 类型(1成员2部门3角色)
        type: integer
    type: object
  inner_req.GetRoleUserIdsReq:
    properties:
      orgId:
        description: 组织id
        type: integer
    type: object
...
  inner_req.GetUserDeptIdsReq:
    properties:
      orgId:
        description: 组织id
        type: integer
      userId:
        description: 人员id
        type: integer
    type: object
...
  inner_req.RoleListByIdsInnerReq:
    properties:
      ids:
        description: Ids ID列表
        items:
          type: integer
        type: array
      orgId:
        description: OrgId 组织ID
        type: integer
    type: object
  inner_req.SimpleReq:
    properties:
      orgId:
        description: 组织id
        type: integer
    type: object
  inner_req.UserAuthorityInnerReq:
    properties:
      orgId:
        description: OrgId 组织ID
        type: integer
      userId:
        description: 用户id
        type: integer
    type: object
  inner_req.UserListByIdsInnerReq:
    properties:
      ids:
        description: Ids ID列表
        items:
          type: integer
        type: array
      orgId:
        description: OrgId 组织ID
        type: integer
    type: object
  inner_resp.DepartmentLeader:
    properties:
      departmentId:
        type: integer
      leaderId:
        type: integer
    type: object
  inner_resp.DeptInfoInnerResp:
    properties:
      code:
        type: string
      id:
        type: integer
      isDelete:
        description: IsDelete
        type: integer
      isHide:
        type: integer
      name:
        type: string
      orgId:
        type: integer
      parentId:
        type: integer
      path:
        type: string
      sort:
        type: integer
      sourceChannel:
        type: string
      status:
        type: integer
    type: object
  inner_resp.GetDeptUserIdsResp:
    properties:
      data:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
    type: object
  inner_resp.GetLeadersByDeptIdsResp:
    properties:
      leaders:
        items:
          $ref: '#/definitions/inner_resp.DepartmentLeader'
        type: array
    type: object
...
  inner_resp.GetMemberSimpleInfoResp:
    properties:
      data:
        items:
          $ref: '#/definitions/inner_resp.SimpleInfo'
        type: array
    type: object
  inner_resp.GetRoleUserIdsResp:
    properties:
      data:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
    type: object
...
  inner_resp.GetUserDeptIdsResp:
    properties:
      deptIds:
        items:
          type: integer
        type: array
    type: object
...
  inner_resp.RepeatMemberInfo:
    properties:
      department:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  inner_resp.RepeatMemberResp:
    properties:
      department:
        items:
          $ref: '#/definitions/inner_resp.RepeatMemberInfo'
        type: array
      role:
        items:
          $ref: '#/definitions/inner_resp.RepeatMemberInfo'
        type: array
      user:
        items:
          $ref: '#/definitions/inner_resp.RepeatMemberInfo'
        type: array
    type: object
  inner_resp.RoleInfoInnerResp:
    properties:
      id:
        type: integer
      name:
        type: string
      orgId:
        type: integer
      roleGroupId:
        type: integer
    type: object
  inner_resp.SimpleInfo:
    properties:
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        description: 父部门id
        type: integer
      status:
        type: integer
    type: object
  inner_resp.UserAuthorityInnerResp:
    properties:
      hasAppPackageOptAuth:
        description: 应用包操作权限
        type: boolean
      hasDeptOptAuth:
        description: 管理部门的权限
        type: boolean
      hasRoleOptAuth:
        description: 角色操作权限
        type: boolean
      isOrgOwner:
        description: 是否是组织拥有者
        type: boolean
      isOutCollaborator:
        description: 是否是外部协作人
        type: boolean
      isSubAdmin:
        description: 是否是子管理员
        type: boolean
      isSysAdmin:
        description: 是否是系统管理员
        type: boolean
      manageAppPackages:
        description: 管理的应用包列表
        items:
          type: integer
        type: array
      manageApps:
        description: ManageApps 管理的应用列表
        items:
          type: integer
        type: array
      manageDepts:
        description: 管理的部门列表
        items:
          type: integer
        type: array
      manageRoles:
        description: 管理的角色列表
        items:
          type: integer
        type: array
      optAuth:
        description: 所拥有操作项
        items:
          type: string
        type: array
      orgId:
        description: 组织ID
        type: integer
      orgSourceChannel:
        description: 组织渠道来源。可能用于后续的权限校验规则。
        type: string
      refDeptIds:
        description: 所管理部门ID列表
        items:
          type: integer
        type: array
      refRoleIds:
        description: 所管理角色ID列表
        items:
          type: integer
        type: array
      userId:
        description: 用户id
        type: integer
    type: object
  inner_resp.UserDeptBindData:
    properties:
      departmentId:
        description: DepartmentId 部门ID
        type: integer
      departmentName:
        description: DepartmentName 部门名称
        type: string
      isLeader:
        description: IsLeader 是否是部门负责人 1是 2否
        type: integer
      outOrgDepartmentCode:
        description: OutOrgDepartmentCode
        type: string
      outOrgDepartmentId:
        description: OutOrgDepartmentId
        type: string
      outOrgDepartmentParentId:
        description: OutOrgDepartmentParentId
        type: string
      positionId:
        description: PositionId 注意此处是org内的职级ID
        type: integer
      positionLevel:
        description: PositionLevel 职级等级
        type: integer
      positionName:
        description: PositionName 职级名称
        type: string
    type: object
  inner_resp.UserInfoInnerResp:
    properties:
      avatar:
        description: Avatar 用户头像
        type: string
      createTime:
        description: CreateTime 创建时间
        type: string
      creator:
        description: Creator 组织成员 创建者
        type: integer
      deptList:
        description: DeptList 部门列表
        items:
          $ref: '#/definitions/inner_resp.UserDeptBindData'
        type: array
      email:
        description: Email 邮箱
        type: string
      id:
        description: Id 用户ID
        type: integer
      isDelete:
        description: IsDelete
        type: integer
      loginName:
        description: LoginName
        type: string
      name:
        description: Name 姓名
        type: string
      namePy:
        description: NamePy 姓名拼音
        type: string
      phoneNumber:
        description: PhoneNumber 手机
        type: string
      phoneRegion:
        description: PhoneRegion 手机区号
        type: string
      roleList:
        description: RoleList 角色列表
        items:
          $ref: '#/definitions/inner_resp.UserRoleBindData'
        type: array
      status:
        description: Status 状态1启用2禁用3离职
        type: integer
      type:
        description: Type 用户类型，1：内部，2：外部
        type: integer
      updateTime:
        description: UpdateTime 创建时间
        type: string
      updator:
        description: Updator 组织成员 最后修改者ID
        type: integer
    type: object
  inner_resp.UserRoleBindData:
    properties:
      roleId:
        type: integer
      roleName:
        type: string
    type: object
...
  resp.CacheUserInfoBo:
    properties:
      corpId:
        type: string
      orgId:
        type: integer
      outUserId:
        type: string
      sourceChannel:
        type: string
      userId:
        type: integer
    type: object
...
info:
  contact: {}
  title: Ying Sou Sou Apis
paths:
...
  /usercenter/inner/api/v1/auth/api-key-auth:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 校验ApiKey接口
      parameters:
      - description: apiKey
        in: query
        name: apiKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.CacheUserInfoBo'
        "400":
          description: ""
      security:
      - token: []
      summary: 校验ApiKey
      tags:
      - 组织
  /usercenter/inner/api/v1/dept/getAllListByIds:
    post:
      consumes:
      - application/json
      description: 根据部门ID列表获取部门信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.DeptListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取所有部门信息列表
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getLeadersByDeptIds:
    post:
      consumes:
      - application/json
      description: 获取部门下的所有leader
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetLeadersByDeptIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inner_resp.GetLeadersByDeptIdsResp'
        "400":
          description: ""
      summary: 获取部门下的所有leader
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getListByIds:
    post:
      consumes:
      - application/json
      description: 根据部门ID列表获取部门信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.DeptListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取部门信息列表（可用）
      tags:
      - 部门（内部调用）
...
  /usercenter/inner/api/v1/dept/getUserDeptIds:
    post:
      consumes:
      - application/json
      description: 获取人员所在的部门id集合
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetUserDeptIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetUserDeptIdsResp'
            type: array
        "400":
          description: ""
      summary: 获取人员所在的部门id集合
      tags:
      - 部门（内部调用）
...
  /usercenter/inner/api/v1/dept/getUserIds:
    post:
      consumes:
      - application/json
      description: 获取部门用户（部门对应用户数组）
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetDeptUserIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetDeptUserIdsResp'
            type: array
        "400":
          description: ""
      summary: 获取部门用户（部门对应用户数组）
      tags:
      - 部门（内部调用）
...
  /usercenter/inner/api/v1/manage-group/add-app:
    post:
      consumes:
      - application/json
      description: 新增应用至改人员所在管理组
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.AddAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 新增应用至人员所在管理组
      tags:
      - 管理组（内部调用）
  /usercenter/inner/api/v1/manage-group/add-pkg:
    post:
      consumes:
      - application/json
      description: 新增应用包至人员所在管理组
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.AddPkgReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 新增应用包至人员所在管理组
      tags:
      - 管理组（内部调用）
...
  /usercenter/inner/api/v1/manage-group/delete-app:
    post:
      consumes:
      - application/json
      description: 从人员所在管理组删除应用
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.DeleteAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 从人员所在管理组删除应用
      tags:
      - 管理组（内部调用）
  /usercenter/inner/api/v1/manage-group/delete-pkg:
    post:
      consumes:
      - application/json
      description: 从人员所在管理组删除应用包
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.DeletePkgReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 从人员所在管理组删除应用包
      tags:
      - 管理组（内部调用）
...
  /usercenter/inner/api/v1/org/add-out-collaborator:
    post:
      consumes:
      - application/json
      description: 新增外部协作人
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.AddOrgOutCollaborator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bo.BaseOrgInfoBo'
        "400":
          description: ""
      summary: 新增外部协作人
      tags:
      - 企业（内部调用）
  /usercenter/inner/api/v1/org/info:
    get:
      consumes:
      - application/json
      description: 获取企业基础信息
      parameters:
      - description: 组织id
        in: query
        name: orgId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bo.BaseOrgInfoBo'
        "400":
          description: ""
      summary: 获取企业基础信息
      tags:
      - 企业（内部调用）
  /usercenter/inner/api/v1/role/getAllListByIds:
    post:
      consumes:
      - application/json
      description: 根据角色ID列表获取角色信息列表（不分状态和是否删除）
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.RoleListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.RoleInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取角色信息列表 （不分状态和是否删除）
      tags:
      - 角色（内部调用）
  /usercenter/inner/api/v1/role/getListByIds:
    post:
      consumes:
      - application/json
      description: 根据角色ID列表获取角色信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.RoleListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.RoleInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取角色信息列表
      tags:
      - 角色（内部调用）
  /usercenter/inner/api/v1/role/getUserIds:
    post:
      consumes:
      - application/json
      description: 获取角色用户（角色对应用户数组）
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetRoleUserIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetRoleUserIdsResp'
            type: array
        "400":
          description: ""
      summary: 获取角色用户（角色对应用户数组）
      tags:
      - 角色（内部调用）
  /usercenter/inner/api/v1/user/auth:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 校验token接口
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.CacheUserInfoBo'
        "400":
          description: ""
      security:
      - token: []
      summary: 校验token
      tags:
      - 认证
  /usercenter/inner/api/v1/user/auth-check-status:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 校验token并验证状态接口
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.CacheUserInfoBo'
        "400":
          description: ""
      security:
      - token: []
      summary: 校验token并验证状态
      tags:
      - 认证
...
  /usercenter/inner/api/v1/user/getAllListByIds:
    post:
      consumes:
      - application/json
      description: 根据成员ID列表获取成员信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.UserListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.UserInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取所有成员信息列表
      tags:
      - 成员（内部调用）
  /usercenter/inner/api/v1/user/getListByIds:
    post:
      consumes:
      - application/json
      description: 根据成员ID列表获取成员信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.UserListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.UserInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取成员信息列表（只获取有效的）
      tags:
      - 成员（内部调用）
  /usercenter/inner/api/v1/user/getMemberSimpleInfo:
    post:
      consumes:
      - application/json
      description: 获取成员简单信息
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetMemberSimpleInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetMemberSimpleInfoResp'
            type: array
        "400":
          description: ""
      summary: 获取成员简单信息
      tags:
      - 成员（内部调用）
  /usercenter/inner/api/v1/user/getRepeatMember:
    post:
      consumes:
      - application/json
      description: 获取重名成员/部门/角色
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.SimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.RepeatMemberResp'
            type: array
        "400":
          description: ""
      summary: 获取重名成员/部门/角色
      tags:
      - 成员（内部调用）
  /usercenter/inner/api/v1/user/getUserAuthority:
    post:
      consumes:
      - application/json
      description: 根据成员ID获取成员权限信息
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.UserAuthorityInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.UserAuthorityInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取成员权限信息
      tags:
      - 成员（内部调用）
swagger: "2.0"